added error handlers and req.user checking

+++ controllers/deleteController.js
@@ -2,7 +2,7 @@ const db = require("../db/queries");
 
 module.exports = {
     post: async (req, res) => {
-        await db.deleteMsg(req.body.id);
+        if (req.user?.admin) await db.deleteMsg(req.body.id);
         res.redirect("/");
     }
 }
